[feat/#31] 티어 화면 식당 id 전달 로직 추가

+++ shared/feature/tier/src/commonMain/kotlin/com/kus/feature/tier/navigation/TierGraph.kt
@@ -19,7 +19,7 @@ data object TierCategorySelect
 fun NavGraphBuilder.tierNavGraph(
     initialProvider: () -> TierFilterState,
     navigateToTierCategorySelect: (TierFilterState) -> Unit,
-    navigateToDetail: () -> Unit,
+    navigateToDetail: (Long) -> Unit,
     popBackStackWithResult: (TierFilterState) -> Unit,
     onBackButtonClick: () -> Unit = {},
 ) {
